chore: move GetBooksAndRelatedTasksController to its own file
