+++ __tests__/unit/presentation/controllers/get-books-and-related-tasks-controller.spec.js
@@ -1,48 +1,14 @@
-const { MissingParamError } = require('../../../../src/utils/errors');
+const { GetBooksAndRelatedTasksController } = require('../../../../src/presentation/controllers');
 const {
   InternalServerError,
   InvalidRequestError,
   NotFoundError,
   ForbiddenError
 } = require('../../../../src/presentation/errors');
-const { HttpResponse } = require('../../../../src/presentation/helpers');
 const {
   ErrorMessagesEnum: { USER_NOT_FOUND }
 } = require('../../../../src/utils/enums');
 const { DataFakerHelper } = require('../../../helpers');
-
-class GetBooksAndRelatedTasksController {
-  constructor({ requestParamsValidator, checkIfRequestIsAllowedService, getBooksAndRelatedTasksService } = {}) {
-    this.requestParamsValidator = requestParamsValidator;
-    this.checkIfRequestIsAllowedService = checkIfRequestIsAllowedService;
-    this.getBooksAndRelatedTasksService = getBooksAndRelatedTasksService;
-  }
-
-  async handler(httpRequest) {
-    try {
-      if (!httpRequest.params) throw new MissingParamError('params');
-
-      const errors = this.requestParamsValidator.validate(httpRequest.params);
-      if (errors) return HttpResponse.badRequest(new InvalidRequestError(errors));
-
-      const isAllowed = this.checkIfRequestIsAllowedService.handler({
-        userId: httpRequest.params.userId,
-        authorization: httpRequest.headers.authorization
-      });
-      if (!isAllowed) return HttpResponse.forbidden();
-
-      const booksAndRelatedTasksModel = await this.getBooksAndRelatedTasksService.handler({
-        userId: httpRequest.params.userId
-      });
-      if (!booksAndRelatedTasksModel) return HttpResponse.notFound(new NotFoundError(USER_NOT_FOUND));
-
-      return HttpResponse.ok(booksAndRelatedTasksModel);
-    } catch (error) {
-      console.log(error);
-      return HttpResponse.exceptionHandler(error);
-    }
-  }
-}
 
 const makeRequestParamsValidatorSpy = () => {
   class RequestParamsValidatorSpy {
