+++ src/presentation/controllers/index.js
@@ -1,5 +1,7 @@
+const GetBooksAndRelatedTasksController = require('./get-books-and-related-tasks-controller');
 const LoginController = require('./login-controller');
 
 module.exports = {
+  GetBooksAndRelatedTasksController,
   LoginController
 };
